Move some napalm module tests to pytest
